feat: TeamAdmin 페이지 라우팅 추가 및 SSLWithPathAtom, SingleSectionLayout 분리
