+++ src/pages/TeamAdmin/TeamAdmin.tsx
@@ -1,0 +1,10 @@
+import { Container } from '@/components/commons/Container';
+import { CommonLayout } from '@/components/layout/CommonLayout';
+
+export const TeamAdmin = () => {
+  return (
+    <CommonLayout>
+      <Container maxW={1090}></Container>
+    </CommonLayout>
+  );
+};
