+++ src/routes/path.tsx
@@ -7,5 +7,6 @@ export const PATH = {
   TEAM_MODIFICATION: '/team-modification',
   JOIN: '/join',
   TEAM_DASHBOARD: '/team-dashboard',
+  TEAM_ADMIN: '/team-admin',
   POSTING: '/posting',
 };
